backend/profile-management/requirements.txt
SQLAlchemy>=2.0
fastapi
strawberry-graphql
uvicorn